requirements.txt
streamlit
openai>=1.0.0
yt-dlp
pydub
openai-whisper
ffmpeg-python